The Crown are here!
Added the new Crown.cat catalogue file with unit definitions and profiles. Updated Alliance.cat to add a conditional 'hidden' modifier for Auxiliary Mine Layer when a specific attachment host is present.

As per usual if you find anything out of place, please reach out to me on the discord, or leave an report on the issue tracker!