Standardize 'Flagship' naming and update generator pool
holy moly its late, I have no time to write a proper summary.
Sultanate are here, fixed have been made across the board, typos and things what didn't work.
again, not everything is up to date with the latest orbat version, but as Union is the only one left, its not far off until i get round to it!
Aiming for all orbats to be completed by the end of the week, and everything updated by the end of the month. (orbat v4 pending)